backend/requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
distro==1.8.0
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
exceptiongroup==1.1.3
fr-core-news-md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.7.0/fr_core_news_md-3.7.0-py3-none-any.whl
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.3
multidict==6.0.4
murmurhash==1.0.10
numpy==1.26.2
oauthlib==3.2.2
openai==0.28.0
packaging==23.2
pandas==2.1.3
preshed==3.0.9
psycopg2==2.9.9
pydantic==2.5.1
pydantic_core==2.14.3
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
thinc==8.2.1
threadpoolctl==3.2.0
tqdm==4.66.1
tweepy==4.14.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wasabi==1.1.2
weasel==0.3.4
yarl==1.9.2
